Properly cast the result of UPDATE queries (even though this result shouldn't be used)

+++ src/Database/Query/UpdateQuery.php
@@ -68,7 +68,7 @@ class UpdateQuery extends AbstractQuery {
 
 		$stmt->execute($this->getParamtersToBind());
 
-		$this->result = Database::getDbh()->lastInsertId();
+		$this->result = (int)Database::getDbh()->lastInsertId();
 
 		self::$totalQueries++;
